requirements.txt
matplotlib==3.10.7
numpy==2.3.5
openai==2.8.1
pandas==2.3.3
Pillow==12.0.0
Requests==2.32.5
scipy==1.16.3
seaborn==0.13.2
setuptools==75.1.0
tiktoken==0.12.0
tqdm==4.66.5
transformers==4.57.3
